mvp/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# OpenAI
openai==1.3.0

# Atlassian integrations
requests==2.31.0
requests-oauthlib==1.3.1

# Data processing
pandas==2.1.4
numpy==1.24.3

# Logging and monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Security
cryptography==41.0.8
pyjwt==2.8.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Performance & Caching
redis==5.0.1
aioredis==2.0.1
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
aiosqlite==0.19.0

# Load Balancing & Monitoring
prometheus-client==0.19.0
gunicorn==21.2.0
psutil==5.9.6
aiohttp==3.9.1

# Database (for local development)
sqlite3
